requirements.txt
diffusers
torch
torchvision
torchaudio
transformers
gradio
accelerate
safetensors
xformers
pillow
numpy
